add toolcalls, with policy control.
