+++ cabal.project
@@ -9,4 +9,4 @@ optional-packages:
 source-repository-package
         type: git
         location: https://github.com/julialongtin/louter.git
-        tag: 6eb00576ed1fd9001659a8c3af52ed3679d2c590
+        tag: 4db253f9ec37d036499ee57c96900ba2be8dc134
